def added

+++ src/test/scala/RepSpec.scala
@@ -10,9 +10,11 @@ class RepSpec extends FlatSpec with ShouldMatchers {
   "A deep embedding of la with Rep types" should "compile" in {
     val x = new VectorDSL {
       def main = {
-        val v1 = DenseVector(liftTerm(1), liftTerm(2), liftTerm(3))
-        val res = (v1 + (DenseVector(liftTerm(3), liftTerm(4), liftTerm(5)) * SparseVector[Int](liftTerm(6), liftTerm(7), liftTerm(8))))
-        val mappedRes = res.map(_ + liftTerm(1))
+        //TODO problem when try to define vector via Map
+//        val v0 = DenseVector(Map(1->liftTerm(1.0), 2->liftTerm(2.0), 3->liftTerm(3.0)))
+        val v1 = DenseVector(liftTerm(1.0), liftTerm(2.0), liftTerm(3.0))
+        val res = (v1 + (DenseVector(liftTerm(3.0), liftTerm(4.0), liftTerm(5.0)) * SparseVector[Double](liftTerm(6.0), liftTerm(7.0), liftTerm(8.0))))
+        val mappedRes = res.map(_ + liftTerm(1.0))
         mappedRes
       }
     }
